Update to Godot 4.3
This also requires explicit casting from float to int in LevelManager

+++ src/TbsTemplate/Scenes/Level/LevelManager.cs
@@ -357,7 +357,7 @@ public partial class LevelManager : Node
                 _target = target;
                 if (!sources.Contains(_path[^1]))
                     PathOverlay.Path = _path = sources.Select((c) => _path.Add(c).Clamp(_selected.Stats.Move)).OrderBy(
-                        (p) => new Vector2I(-p[^1].DistanceTo(cell), p[^1].DistanceTo(_path[^1])),
+                        (p) => new Vector2I(-(int)p[^1].DistanceTo(cell), (int)p[^1].DistanceTo(_path[^1])),
                         static (a, b) => a < b ? -1 : a > b ? 1 : 0
                     ).First();
             }
